requirements.txt
langchain-community
langchain-experimental
langchain-openai
python-dotenv
openai
psycopg2-binary
sqlalchemy
pyodbc